% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{find_shift_weights}
\alias{find_shift_weights}
\title{Find optimal weight shift for sensitivity bounds}
\usage{
find_shift_weights(inf, gamma, y, wts = NULL)
}
\description{
Find optimal weight shift for sensitivity bounds
}
\keyword{internal}
