% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_utils.R
\name{parse_psbart_formula}
\alias{parse_psbart_formula}
\title{Parse Principal Stratification Formula}
\usage{
parse_psbart_formula(formula, data)
}
\arguments{
\item{formula}{A formula with three parts separated by |}

\item{data}{A data.frame containing the variables}
}
\value{
A list with components X, X_raw, Y, Z, W
}
\description{
Parses a formula in 2SLS style: Y ~ X1 + X2 | Z | W
}
\keyword{internal}
